requirements.txt
Pillow~=8.1.2
pytesseract==0.3.8
